+++ GrosRobot/Arduino/gros_robot_full_control_v3/gros_robot_full_control_actionneurs_v3/Capteur_ultrason.h
@@ -3,7 +3,7 @@ class Capteur_ultrason
     public:
     Capteur_ultrason(int pin_trig,int pin_echo);
     void set_capteur();
-    int get_distance();  
+    float get_distance();  
 
     private:
     int m_pin_trig;
